Move celery beat tasks into `tasks` directory
